requirements.txt
# Core Framework
flask==2.3.3
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0

# HTTP and Web
requests==2.31.0
aiohttp==3.8.6

# Image Processing and Computer Vision
pillow==10.0.1
opencv-python==4.8.1.78
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.200  # YOLO
albumentations==1.3.1

# Scientific Computing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.3

# Geospatial Analysis
geopy==2.3.0
geopandas==0.14.0
shapely==2.0.2
rasterio==1.3.9
folium==0.14.0
contextily==1.4.0
osmnx==1.6.0

# Earth Engine and Satellite Data
earthengine-api==0.1.376
google-cloud-storage==2.10.0
planetary-computer==0.4.9
pystac-client==0.7.5

# AI and Machine Learning
openai==0.28.1
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.1.0

# Data Processing
python-dotenv==1.0.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.7

# Visualization and Plotting
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Development and Testing
jupyter==1.0.0
pytest==7.4.2
pytest-asyncio==0.21.1
black==23.7.0
flake8==6.0.0
mypy==1.6.1

# Production
gunicorn==21.2.0
redis==5.0.1
celery==5.3.3

# Monitoring and Logging
prometheus-client==0.18.0
structlog==23.2.0
